Rename Branch to Family and rename all related functions and variables
"Branch" as a concept is really just about the location in 3D space where the
vertical connector descending from two parents splits into separate
connectors to each child. This loosely was used to mean "the relationships
between nodes representing a nuclear family", but that was less than clear,
so this change clarifies the terminology.