Change unit tests from Fable.Mocha on node.js to Expecto on .NET
The Fable unit testing ecosystem seems pretty immature. For example, basic
assertions are broken in Fable.Mocha. Expected & actual values are not output
which is a pretty big miss. Also, FsCheck is not supported in Fable. This is
going to be a problem down the road, so I'm cutting my losses and moving the
tests to .NET now.