Add FamilyGraph to the Model and change NodeId to PersonId
This is a pretty big change that will set us up for the next few steps in the
evolution of this app. FamilyGraph is the module that will handle all
interrogation of the family "tree" data structure (which is really an
undirected multi-graph rather than a tree). This change adds logic to create
the graph from a plausibly tabular form of input data, and calculate the
Family records needed for rendering from that.

Changing NodeId to PersonId makes sense when you assume that every person in
the graph will have a rendered node, and that relationships in the graph will
be implemented by index rather than direct references (it's just easier that
way).