Move Family back to ViewModel and add IGraphViewFactory & its implementation
Family is really rendering-specific, so it doesn't belong in Model. This
moves it back to ViewModel, along with the functionality that extracts the
collection of Family records from a FamilyGraph. This is now encapsulated in
a new interface, IGraphViewFactory, which is called by the top-level App.

The goal is to slowly replace hard-coded data with dynamically-computed data.
Ultimately the source person/parent data will come from an external file.

The next step is to implement IGraphViewFactory.LayoutTreeNodes, which is
currently still assuming a hard-coded FamilyGraph.