Reintroducing NodeId to accomodate 1:M relationship from Person to TreeNode
It turns out it was a bad idea to get rid of NodeId. The issue is how to
represent co-parent relationships when multiple huwilp are rendered. If every
person only gets one node, but each node is visually aligned with its Wilp
tree, and if children branch off the middle of the co-parent connector, then
we get a weird looking visualization that no longer keeps Wilp members in
trees. The solution will be to have "proxy" nodes for each such co-parent
where necessary. This requires having an explicit NodeId as well as a way to
unambiguously map from a PersonId to a NodeId in the context of a rendered
Wilp tree. These concepts will be further refined and tested in the future.
The goal for now is to have just enough complexity in the ViewModel to enable
implementation of dynamic layout.