Implement animation for layout and undo/redo
This change adds support for animation to TreeNodeMesh and its corresponding
ViewModel logic. This required introducing the concept of a TargetPosition
for each node, as well as an IsAnimating flag. The actual animation loop is
driven from useFrame in TreeNodeMesh.

While this change works, it's a bit awkward since undo/redo had to be hacked
up to not quite actually undo/redo, but instead set target positions. If
dragging isn't disabled during animation, weird things can happen, like all
node positions being saved during automatic layout. This suggests we need a
conceptual re-think of how undo/redo work in the presence of animation.

This app is starting to feel more like a game or simulation now that it
supports animation, and I'm seriously considering a major refactoring to use
the ECS (Entity Component System) architecture instead of, or perhaps in
addition to, MVU (Model View Update). This might have better separation of
concerns, and I'm hoping it would make it easier to reason about how
different behaviours interact.