Disable Undo and Redo when nodes are animating
This fixes a bug where clicking Undo or Redo while in the middle of animating
a previous Undo or Redo would lose the current state.

This is really a deeper issue with how undo state is represented, but fixing
that is a bigger undertaking. For now, just make the problematic behaviour
impossible in the UI.