Simplified and clarified drag logic
This change is based on observation that the origin provided by DragControls
is always (0, 0, 0) and therefore serves no useful purpose. It also turns out
that "offset" and the co-ordinates for DragBy (formerly known as DragTo) were
being confused in terms of which one is relative and which one is absolute.
Offset has been renamed to Origin to reflect that it is the absolute one.
Apparently the local matrix transform we do in handleDrag produces a local
delta from the initial drag position. Variable names have been clarified so
the drag logic in ViewModel is easier to follow.