Implement dynamic node layout
After this change, tree node positions no longer need to be hard-coded.

The core of the layout algorithm is the LayoutBox module, which implements an
"almost rectangle" abstraction ("almost" because it can have notched corners
in the top left and right) that nests intelligently. Leaf and composite boxes
compose by attaching them horizontally or vertically. Units of Measure are
used to keep the different co-ordinate systems straight while combining boxes
and there are operations to "reframe" vectors or boxes to a different unit as
they're being composed up the family tree.

The layout algorithm relies on the Wilp Tree catamorphism that previously
snuck its way into main (it should have been a separate commit from the
BirthOrder change).

The layout algorithm is packaged up in a new module called Scene, which also
handles other initialization tasks such as extracting families (the type of
which is now called RenderedFamily) from the FamilyGraph.

To better organize this large change, all ViewModel related files have been
moved to a ViewModel sub-folder and TypeScript imports have been updated
accordingly.