Abstract parameter type of Scene.extractFamilies
This is a clairvoyant commit to make future changes easier to review. It
changes the parameter sequence type of Scene.extractFamilies from TreeNode to
a generic type parameter based on an interface. This is because the TreeNode
type will disappear in the future as we move from a pure MVU core to one
based on the ECS (Entity Component System) architecture.